Correção aplicada post http://www.projetoacbr.com.br/forum/topic/29049-problemas-no-acbrdownload/?do=findComment&comment=192070
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@11556 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrTCP/ACBrDownloadClass.pas
@@ -336,11 +336,13 @@ begin
             if (fResultCode = 200) or (fResultCode = 206) or (fResultCode = 226) then
             begin
               sFileName := Copy(fFilePart, 1, Length(fFilePart) -5);
-              if FileExists(sFileName) then
+              if FileExists(sFileName) and FilesExists(fFilePart) then
                 if not DeleteFile(sFileName) then
                   raise Exception.Create(Format('Não foi possível excluir o arquivo: %s', [sFileName]));
-              if not RenameFile(fFilePart, sFileName) then
-                raise Exception.Create(Format('Não foi possível renomear o arquivo: %s para %s', [fFilePart, sFileName]));
+
+              if FilesExists(fFilePart) then              
+                if not RenameFile(fFilePart, sFileName) then
+                  raise Exception.Create(Format('Não foi possível renomear o arquivo: %s para %s', [fFilePart, sFileName]));
             end;
          end
        end;
